game over ui and high score tracking

+++ Assets/Scripts/ClusterManager.cs
@@ -8,7 +8,7 @@ public class ClusterManager : MonoBehaviour
 
     //4 difficulties [1, 4]
     public GameObject SpawnCluster(int difficulty, float spawnLoc){
-        int i = Random.Range(0, difficulty*origClusters.Length/4);
+        int i = Random.Range(0, (int) ((float) difficulty*(float) (origClusters.Length)/4.0f));
         GameObject cluster = Instantiate(origClusters[i] , new Vector2(0f, spawnLoc), Quaternion.identity);
         return cluster;
     }
